fix crashing
