+++ src/main/java/collection/PaymentInfo.java
@@ -1,7 +1,7 @@
-package com.example.BookstoreBakersButBetter.collection;
+package collection;
 
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
+//import jakarta.persistence.GeneratedValue;
+//import jakarta.persistence.GenerationType;
 import lombok.Builder;
 import lombok.Data;
 import org.springframework.data.annotation.Id;
@@ -12,7 +12,7 @@ import org.springframework.data.annotation.Id;
 public class PaymentInfo {
 
     @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
+  //  @GeneratedValue(strategy = GenerationType.AUTO)
     private String paymentInfoId; //primary id
   //  @Id (should this be used here?)
     private String addressID; //foreign key of the address associated with the card
