+++ src/main/java/com/example/BookstoreBakersButBetter/BookstoreBakersButBetterApplication.java
@@ -2,8 +2,10 @@ package com.example.BookstoreBakersButBetter;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
 
 @SpringBootApplication
+@EnableMongoRepositories
 public class BookstoreBakersButBetterApplication {
 
 	public static void main(String[] args) {
