--- src/main/java/com/example/BookstoreBakersButBetter/config/MongoSSLConfiguration.java
@@ -1,4 +1,0 @@
-package com.example.BookstoreBakersButBetter.config;
-
-public class MongoSSLConfiguration {
-}
