--- src/main/java/com/example/BookstoreBakersButBetter/controller/RequestBody.java
@@ -1,4 +1,0 @@
-package com.example.BookstoreBakersButBetter.controller;
-
-public @interface RequestBody {
-}
