--- src/main/java/com/example/BookstoreBakersButBetter/controller/UserController.java
@@ -1,4 +1,0 @@
-package com.example.BookstoreBakersButBetter.controller;
-
-public class UserController {
-}
