--- src/main/java/com/example/BookstoreBakersButBetter/repository/BookRepository.java
@@ -1,20 +1,0 @@
-package com.example.BookstoreBakersButBetter.repository;
-
-import com.example.BookstoreBakersButBetter.collection.book;
-import org.springframework.data.mongodb.repository.MongoRepository;
-import org.springframework.stereotype.Repository;
-import org.springframework.data.domain.Example;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
-import org.springframework.data.mongodb.repository.MongoRepository;
-import org.springframework.data.repository.query.FluentQuery;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-import java.util.Optional;
-import java.util.function.Function;
-
-@Repository
-public interface BookRepository extends MongoRepository<book,String> {
-}
