--- src/main/java/com/example/BookstoreBakersButBetter/repository/FavoritesListRepository.java
@@ -1,4 +1,0 @@
-package com.example.BookstoreBakersButBetter.repository;
-
-public interface FavoritesListRepository extends JpaRespository<> {
-}
