--- src/main/java/com/example/BookstoreBakersButBetter/repository/UserRepository.java
@@ -1,7 +1,0 @@
-package com.example.BookstoreBakersButBetter.repository;
-
-import org.springframework.boot.autoconfigure.security.SecurityProperties;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<SecurityProperties.User, Integer> {
-}
