+++ src/main/java/controller/RequestBody.java
@@ -1,0 +1,4 @@
+package controller;
+
+public @interface RequestBody {
+}
