+++ src/main/java/controller/UserController.java
@@ -1,0 +1,7 @@
+package controller;
+
+import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
+
+@EnableMongoRepositories
+public class UserController {
+}
