+++ src/main/java/controller/bookController.java
@@ -1,21 +1,24 @@
-package com.example.BookstoreBakersButBetter.controller;
+package controller;
 
-import com.example.BookstoreBakersButBetter.collection.book;
-import com.example.BookstoreBakersButBetter.service.BookService;
+import collection.book;
+import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
+import service.BookService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/book")
+@EnableMongoRepositories
 public class bookController {
     @Autowired
     private BookService bookService;
 
     @Override
     public String save(@RequestBody book book){
-        return bookService.save(book).getbookId();
-
+     //   return bookService.save(book).getbookId();
+        //return null;
+        return "";
     }
 
 }
