+++ src/main/java/model/favoritesList.java
@@ -1,4 +1,4 @@
-package com.example.BookstoreBakersButBetter.model;
+package model;
 
 import lombok.Getter;
 import lombok.Setter;
