+++ src/main/java/repository/BookRepository.java
@@ -1,0 +1,9 @@
+package repository;
+
+import collection.book;
+import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface BookRepository extends MongoRepository<book,String> {
+}
