+++ src/main/java/repository/FavoritesListRepository.java
@@ -1,0 +1,4 @@
+package repository;
+
+public interface FavoritesListRepository {
+}
