+++ src/main/java/repository/UserRepository.java
@@ -1,0 +1,7 @@
+package repository;
+
+//import org.springframework.boot.autoconfigure.security.SecurityProperties;
+//import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository {
+}
