+++ src/main/java/service/BookService.java
@@ -1,0 +1,7 @@
+package service;
+
+import collection.book;
+
+public interface BookService {
+    Object save(book book);
+}
