+++ src/main/java/service/UserService.java
@@ -1,4 +1,4 @@
-package com.example.BookstoreBakersButBetter.service;
+package service;
 
 import org.apache.catalina.User;
 import org.springframework.stereotype.Service;
