+++ src/main/resources/application.properties
@@ -8,9 +8,10 @@ spring.servlet.multipart.max-file-size=10MB
 #specificies max size allowed for multipart/form-data requests
 spring.servlet.multipart.max-request-size=10MB
 ##--file upload stuff--##
-spring.data.mongodb.authentication-database=admin
-spring.data.mongodb.username=baker1
-spring.data.mongodb.password=bread1
+#spring.data.mongodb.authentication-database=admin
+#spring.data.mongodb.username=baker1
+#spring.data.mongodb.password=bread1
+spring.data.mongodb.uri=mongodb+srv://baker1:bread1@bookstorebakersbutbette.pvz6mfb.mongodb.net/
 spring.data.mongodb.database=bookstoreBakersButBetter
-spring.data.mongodb.port=27017
-spring.data.mongodb.host=localhost
+#spring.data.mongodb.port=27017
+#spring.data.mongodb.host=localhost
